fix : Fix exceed time limit problem
In the previous version, I implemented a function "getMaxVolume(int nowVolume, int i")
which returns the maximum volume possible
when the starting volume is "nowVolume"
and the index of the arry V[] starts from integer "i".
In this new version, I implemented a function "getVolumeChange(int i, int sumOfVolumeChange)"
which returns the maximum volume change possible
when the sum of previous volume changes is "sumOfVolumeChange"
and the index of the arry V[] starts from integer "i".

Honestly, don't know why the exceeded time limit problem was solved
because both version use the same cache size 100 * 1000.