fix : Fix Exceed time limit problem
Added "const integer MIN = -1001"
to check the cases when changing the volume is impossible.
Since the cache was initialized to -1,
the previous version of the code was impossible to determine
if the cache value was uncalculated or was saved as an impossible case.
That was the reason for time exceeded problem for this version
eventhough it used the same size of the cache